+++ driver/lib/sessions/session.ts
@@ -7,12 +7,14 @@ import {
   startIOSSession, connectIOSSession
 } from './ios';
 import { PLATFORM } from '../platform';
+import type { XCUITestDriver } from 'appium-xcuitest-driver';
+import type {AndroidUiautomator2Driver} from 'appium-uiautomator2-driver';
 
 
 export const reConnectFlutterDriver = async function(this: FlutterDriver, caps: Record<string, any>) {
   // setup proxies - if platformName is not empty, make it less case sensitive
   if (!caps.platformName) {
-    this.log.errorAndThrow(`No platformName was given`);
+    this.log.errorWithException(new Error(`No platformName was given`));
   }
 
   switch (_.toLower(caps.platformName)) {
@@ -23,9 +25,11 @@ export const reConnectFlutterDriver = async function(this: FlutterDriver, caps: 
       this.socket = await connectAndroidSession.bind(this)(this.proxydriver, caps, true);
       break;
     default:
-      this.log.errorAndThrow(
-        `Unsupported platformName: ${caps.platformName}. ` +
-        `Only the following platforms are supported: ${_.keys(PLATFORM)}`
+      this.log.errorWithException(
+        new Error(
+          `Unsupported platformName: ${caps.platformName}. ` +
+          `Only the following platforms are supported: ${_.keys(PLATFORM)}`
+        )
       );
   }
 };
@@ -36,21 +40,23 @@ export const createSession: any = async function(this: FlutterDriver, sessionId:
     switch (_.toLower(caps.platformName)) {
       case PLATFORM.IOS:
         [this.proxydriver, this.socket] = await startIOSSession.bind(this)(caps, ...args);
-        this.proxydriver.relaxedSecurityEnabled = this.relaxedSecurityEnabled;
-        this.proxydriver.denyInsecure = this.denyInsecure;
-        this.proxydriver.allowInsecure = this.allowInsecure;
+        (this.proxydriver as XCUITestDriver).relaxedSecurityEnabled = this.relaxedSecurityEnabled;
+        (this.proxydriver as XCUITestDriver).denyInsecure = this.denyInsecure;
+        (this.proxydriver as XCUITestDriver).allowInsecure = this.allowInsecure;
 
         break;
       case PLATFORM.ANDROID:
         [this.proxydriver, this.socket] = await startAndroidSession.bind(this)(caps, ...args);
-        this.proxydriver.relaxedSecurityEnabled = this.relaxedSecurityEnabled;
-        this.proxydriver.denyInsecure = this.denyInsecure;
-        this.proxydriver.allowInsecure = this.allowInsecure;
+        (this.proxydriver as AndroidUiautomator2Driver).relaxedSecurityEnabled = this.relaxedSecurityEnabled;
+        (this.proxydriver as AndroidUiautomator2Driver).denyInsecure = this.denyInsecure;
+        (this.proxydriver as AndroidUiautomator2Driver).allowInsecure = this.allowInsecure;
         break;
       default:
-        this.log.errorAndThrow(
-          `Unsupported platformName: ${caps.platformName}. ` +
-          `Only the following platforms are supported: ${_.keys(PLATFORM)}`
+        this.log.errorWithException(
+          new Error(
+            `Unsupported platformName: ${caps.platformName}. ` +
+            `Only the following platforms are supported: ${_.keys(PLATFORM)}`
+          )
         );
     }